\name{FDiff}
\alias{FDiff}
\title{creates a variable to be tracked by the computation of the Jacobian}
\usage{
  FDiff(x, name)
}
\arguments{
  \item{x}{a vector of current values}

  \item{name}{a anme that uniquly identify this variable in
  the overall}
}
\description{
  creates a variable to be tracked by the computation of
  the Jacobian
}

